Update - I Own the Keyboard Edition
Actually got a corsair k70 RGB so I can test and debug as I develop. The
Bomb ticking was improved, the ammo counting for weapons works, and the
health and armor bars work. The keyboard now also lights all keys grey
when not in a game/on a team. Other minor cleanup.